refactor: move change stream cursor into separate file
