+++ src/index.ts
@@ -29,6 +29,7 @@ export {
   ObjectId,
   Timestamp
 } from './bson';
+export { ChangeStreamCursor } from './cursor/change_stream_cursor';
 /**
  * @public
  * @deprecated Please use `ObjectId`
@@ -174,8 +175,6 @@ export type {
   ChangeStreamCollModDocument,
   ChangeStreamCreateDocument,
   ChangeStreamCreateIndexDocument,
-  ChangeStreamCursor,
-  ChangeStreamCursorOptions,
   ChangeStreamDeleteDocument,
   ChangeStreamDocument,
   ChangeStreamDocumentCollectionUUID,
@@ -259,6 +258,7 @@ export type {
 } from './cursor/abstract_cursor';
 export type { InternalAbstractCursorOptions } from './cursor/abstract_cursor';
 export type { AggregationCursorOptions } from './cursor/aggregation_cursor';
+export type { ChangeStreamCursorOptions } from './cursor/change_stream_cursor';
 export type { DbOptions, DbPrivate } from './db';
 export type { AutoEncrypter, AutoEncryptionOptions, AutoEncryptionTlsOptions } from './deps';
 export type { Encrypter, EncrypterOptions } from './encrypter';
